Create CharacterScript
So far the script allows the character to go to a random forest to search for either food or resources (the player commands them using buttons). The character is deselected when clicked away from or in the middle of doing an action and can only be selected when they are not busy. At this point of time I still need to write the Breed coroutine which will allow the character to breed with other characters.