Create Panel Hover Script
This is a small script I made to detect whether or not the player has their mouse hovered over the panel at the bottom of the screen which appears when in build mode or when a character is selected. I made this so that the character which is selected isn't deselected when the panel or a button is pressed. The reason this was important to create is because due to some code in the character control script, when a button is pressed it would actually deselect the character and not do anything because the player would be clicking without hovering over the location of the character, so this script was necessary.