Create Breed Script
In the breed script, the selected character is the character which the player originally selected and chose to breed with. The selected breeding character is the character which the player selects whilst the selected character is breeding. When two partners are selected, the original selected character will go towards their partner and a child will be instantiated.