Create Blueprint Script
The blueprint script shows a blueprint of the selected building that the player wants to place (which is essentially a see through version of the building). It will only appear if the player has enough resources to place the building in the first place. The blueprint will follow the player's mouse until they press the left mouse button, and then the building will be placed.