Day 1 - character doidling
I am trying to add "doidling" for my character. What this means is they will walk about in the general area of the camp to make the game feel a bit more lively. Unfortunately it isn't working quite yet as it is showing an error stating that the Update function can't take parameters, however I believe that I am very close to being finished.