Update Character Script
The buildings and doidle area have now been set to triggers and so onCollisionEnter has been changed accordingly. The baby is also now born on the character script when the selected character and breeding partner collide, I did this because it used to only happen when the two characters' transforms were the same but this isn't possible now that they both have rigidbodies. The character isn't doidling propely yet and just ends up teleporting at a weird position outside of the doidle area once collision starts, I feel that this may be because the coroutine for doidling isn't started from Update, I will look for a fix next time.